@@ -3,21 +3,34 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build"
+    "dev": "vite",
+    "build": "run-p type-check \"build-only {@}\" --",
+    "preview": "vite preview",
+    "build-only": "vite build",
+    "type-check": "vue-tsc --build",
+    "lint": "eslint . --fix"
   },
   "dependencies": {
+    "@vue/compiler-sfc": "^3.3.4",
     "core-js": "^3.6.5",
-    "pug": "^3.0.0",
-    "pug-plain-loader": "^1.0.0",
-    "vue": "^2.6.11",
-    "vue-router": "^3.2.0",
-    "vue-select": "^3.20.0"
+    "vue": "^3.5.13",
+    "vue-router": "^4.5.0",
+    "vue3-select-component": "^0.14.0"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "^4.1.1",
-    "@vue/cli-plugin-router": "^4.1.1",
-    "@vue/cli-service": "^4.1.1",
-    "vue-template-compiler": "^2.6.11"
+    "@tsconfig/node22": "^22.0.1",
+    "@types/file-saver": "^2.0.7",
+    "@types/node": "^22.14.1",
+    "@vitejs/plugin-vue": "^5.2.3",
+    "@vue/eslint-config-typescript": "^14.5.0",
+    "@vue/tsconfig": "^0.7.0",
+    "eslint": "^9.22.0",
+    "eslint-plugin-vue": "~10.0.0",
+    "jiti": "^2.4.2",
+    "npm-run-all2": "^7.0.2",
+    "typescript": "~5.8.0",
+    "vite": "^6.3.6",
+    "vite-plugin-vue-devtools": "^7.7.2",
+    "vue-tsc": "^2.2.8"
   }
 }
