@@ -1,73 +1,84 @@
-<template lang="pug">
-.root
-  .container
-    .title-container(@click="toHome")
-      img.cat-image(:src="require('../assets/cat.png')")
-      h1.title-label SER225
-    .routes-container
-      .router-link-container
-        p.router-link(@click="toHome" :style="homeRouteStyle") Home
-      .router-link-container
-        p.router-link(@click="toTeams" :style="teamsRouteStyle") Teams
-      .router-link-container
-        p.router-link(@click="toSearch" :style="searchRouteStyle") Search
-</template>
-
-<script>
+<script setup lang="ts">
+import { ref, computed, onMounted, onBeforeUnmount } from 'vue'
+import { useRoute, useRouter } from 'vue-router'
 import { eventBus } from '../helpers/EventBus.js'
 
-export default {
-  name: 'Navbar',
-  components: {
-  },
-  data() {
-    return {
-      selectedRoute: ''
-    }
-  },
-  created() {
-    this.updateSelectedRoute(this.$route.name)
-    eventBus.$on('route-change', (selectedRoute) => {
-      this.updateSelectedRoute(selectedRoute)
-    })
-  },
-  computed: {
-    homeRouteStyle() {
-      return this.selectedRoute === 'Home' ? { 'border-bottom': '2px solid #253C59' } : {}
-    },
-    teamsRouteStyle() {
-      return this.selectedRoute === 'Teams' ? { 'border-bottom': '2px solid #253C59' } : {}
-    },
-    searchRouteStyle() {
-      return this.selectedRoute === 'Search' ? { 'border-bottom': '2px solid #253C59' } : {}
-    },
-  },
-  methods: {
-    updateSelectedRoute(selectedRoute) {
-      this.selectedRoute = selectedRoute
-      if (this.selectedRoute === 'Team') {
-        this.selectedRoute = 'Teams'
-      }
-    },
-    toHome() {
-      return this.$router.push('/').catch(() => {})
-    },
-    toTeams() {
-      return this.$router.push('/teams').catch(() => {})
-    },
-    toSearch() {
-      return this.$router.push('/search').catch(() => {})
-    },
-  }
+// Image import (Vue 3 compatible)
+import catImg from '../assets/cat.png'
+
+const router = useRouter()
+const route = useRoute()
+
+const selectedRoute = ref<string>('')
+
+const updateSelectedRoute = (r: string) => {
+  selectedRoute.value = r === 'Team' ? 'Teams' : r
 }
+
+onMounted(() => {
+  updateSelectedRoute(route.name as string)
+
+  eventBus.on('route-change', updateSelectedRoute)
+})
+
+onBeforeUnmount(() => {
+  eventBus.off('route-change', updateSelectedRoute)
+})
+
+const homeRouteStyle = computed(() =>
+  selectedRoute.value === 'Home'
+    ? { borderBottom: '2px solid #253C59' }
+    : {}
+)
+
+const teamsRouteStyle = computed(() =>
+  selectedRoute.value === 'Teams'
+    ? { borderBottom: '2px solid #253C59' }
+    : {}
+)
+
+const searchRouteStyle = computed(() =>
+  selectedRoute.value === 'Search'
+    ? { borderBottom: '2px solid #253C59' }
+    : {}
+)
+
+const toHome = () => router.push('/').catch(() => {})
+const toTeams = () => router.push('/teams').catch(() => {})
+const toSearch = () => router.push('/search').catch(() => {})
 </script>
+
+<template>
+  <div class="root">
+    <div class="container">
+      <div class="title-container" @click="toHome">
+        <img class="cat-image" :src="catImg" />
+        <h1 class="title-label">SER225</h1>
+      </div>
+
+      <div class="routes-container">
+        <div class="router-link-container">
+          <p class="router-link" @click="toHome" :style="homeRouteStyle">Home</p>
+        </div>
+
+        <div class="router-link-container">
+          <p class="router-link" @click="toTeams" :style="teamsRouteStyle">Teams</p>
+        </div>
+
+        <div class="router-link-container">
+          <p class="router-link" @click="toSearch" :style="searchRouteStyle">Search</p>
+        </div>
+      </div>
+    </div>
+  </div>
+</template>
 
 <style scoped>
 .container {
   padding-left: 20px;
-  background: #99B4BF;
+  background: #99b4bf;
   height: 55px;
-  border-bottom: 1px solid #253C59;
+  border-bottom: 1px solid #253c59;
   display: flex;
   flex-direction: row;
   align-items: center;
@@ -104,7 +115,7 @@ export default {
 }
 
 .router-link {
-  color: #253C59;
+  color: #253c59;
   margin: 0px;
   padding-left: 0px;
   padding-right: 0px;
@@ -113,7 +124,7 @@ export default {
 
 .router-link:hover {
   cursor: pointer;
-  border-bottom: 2px solid #253C59;
+  border-bottom: 2px solid #253c59;
 }
 
 .router-link::selection {
