@@ -1,22 +1,6 @@
-<template lang="pug">
-.root
-  ThreeTwo(v-if="team === '3-2'")
-  Art(v-else-if="team === 'art'")
-  DevDynasty(v-else-if="team === 'devdynasty'")
-  DreamTeam2(v-else-if="team === 'dream-team-2.0'")
-  ElementalHero(v-else-if="team === 'elemental-hero'")
-  Error101(v-else-if="team === 'error101'")
-  FireFox(v-else-if="team === 'firefox'")
-  LetUsCook(v-else-if="team === 'let-us-cook'")
-  OnlyGit(v-else-if="team === 'onlygit'")
-  TeamPlasma(v-else-if="team === 'team-plasma'")
-  XoxoGossipGirl(v-else-if="team === 'xoxo-gossip-girl'")
+<script setup lang="ts">
+import { defineProps } from 'vue'
 
-  template(v-else)
-    p.error-message Invalid Team!
-</template>
-
-<script>
 import ThreeTwo from './fall2023/3-2/ThreeTwo.vue'
 import Art from './fall2023/art/Art.vue'
 import DevDynasty from './fall2023/devdynasty/DevDynasty.vue'
@@ -29,29 +13,28 @@ import OnlyGit from './fall2023/onlygit/OnlyGit.vue'
 import TeamPlasma from './fall2023/team-plasma/TeamPlasma.vue'
 import XoxoGossipGirl from './fall2023/xoxo-gossip-girl/XoxoGossipGirl.vue'
 
-export default {
-  name: 'Fall2023',
-  components: {
-    ThreeTwo,
-    Art,
-    DevDynasty,
-    DreamTeam2,
-    ElementalHero,
-    Error101,
-    FireFox,
-    LetUsCook,
-    OnlyGit,
-    TeamPlasma,
-    XoxoGossipGirl
-  },
-  props: {
-    team: {
-      type: String,
-      required: true
-    }
-  }
-}
+const props = defineProps<{
+  team: string
+}>()
 </script>
+
+<template>
+  <div class="root">
+    <ThreeTwo v-if="props.team === '3-2'" />
+    <Art v-else-if="props.team === 'art'" />
+    <DevDynasty v-else-if="props.team === 'devdynasty'" />
+    <DreamTeam2 v-else-if="props.team === 'dream-team-2.0'" />
+    <ElementalHero v-else-if="props.team === 'elemental-hero'" />
+    <Error101 v-else-if="props.team === 'error101'" />
+    <FireFox v-else-if="props.team === 'firefox'" />
+    <LetUsCook v-else-if="props.team === 'let-us-cook'" />
+    <OnlyGit v-else-if="props.team === 'onlygit'" />
+    <TeamPlasma v-else-if="props.team === 'team-plasma'" />
+    <XoxoGossipGirl v-else-if="props.team === 'xoxo-gossip-girl'" />
+
+    <p v-else class="error-message">Invalid Team!</p>
+  </div>
+</template>
 
 <style scoped>
 .error-message {
